Move Responses provider event handling to stream events
